Move base tables to `query` package
