+++ rdbms/src/main/java/io/spine/server/storage/jdbc/query/WriteQueryFactory.java
@@ -22,7 +22,7 @@ package io.spine.server.storage.jdbc.query;
 /**
  * An interface of a {@linkplain StorageQuery query} factory for the write queries.
  *
- * <p>Each JDBC {@linkplain io.spine.server.storage.jdbc.AbstractTable table} uses a query factory
+ * <p>Each JDBC {@linkplain AbstractTable table} uses a query factory
  * to construct its SQL queries to the database.
  *
  * @param <I> type of the ID of the record
